requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
py-cord==2.6.1
python-dotenv==1.1.1
regex==2025.9.18
typing_extensions==4.15.0
yarl==1.22.0
